Fixed AppBar Animation
Closes https://github.com/YuganshT79/Whatsapp-Copy-Layout/issues/1
