+++ app/src/main/java/com/yuganshtyagi/whatsapp_copy/Adapter/StatusAdapter.java
@@ -1,4 +1,4 @@
-package com.yuganshtyagi.whatsapp_copy;
+package com.yuganshtyagi.whatsapp_copy.Adapter;
 
 import android.content.Context;
 import android.support.annotation.NonNull;
@@ -10,6 +10,7 @@ import android.widget.ImageView;
 import android.widget.TextView;
 
 import com.bumptech.glide.Glide;
+import com.yuganshtyagi.whatsapp_copy.R;
 
 /**
  * Created by Yugansh Tyagi on 3/25/2018.
