+++ app/src/main/java/com/yuganshtyagi/whatsapp_copy/Fragment/ChatsFragment.java
@@ -1,4 +1,4 @@
-package com.yuganshtyagi.whatsapp_copy;
+package com.yuganshtyagi.whatsapp_copy.Fragment;
 
 import android.graphics.Color;
 import android.os.Bundle;
@@ -11,28 +11,31 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.yuganshtyagi.whatsapp_copy.Adapter.ChatAdapter;
+import com.yuganshtyagi.whatsapp_copy.R;
+import com.yuganshtyagi.whatsapp_copy.Utils.SeparatorDecoration;
+
 /**
  * Created by Yugansh Tyagi on 3/21/2018.
  */
 
-public class StatusFragment extends Fragment {
+public class ChatsFragment extends Fragment {
 
     RecyclerView recyclerView;
 
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.status_fragment_layout, null, false);
-
+        View view = inflater.inflate(R.layout.chats_fragment_layout, container, false);
 
         SeparatorDecoration decoration = new SeparatorDecoration(
                 getContext(),
                 Color.parseColor("#EAEAEA"),
                 0.5f);
-        recyclerView = view.findViewById(R.id.status_rv);
+        recyclerView = view.findViewById(R.id.chat_rv);
         recyclerView.setHasFixedSize(true);
         recyclerView.setLayoutManager(new LinearLayoutManager(getContext()));
-        recyclerView.setAdapter(new StatusAdapter(getContext()));
+        recyclerView.setAdapter(new ChatAdapter(getContext()));
         recyclerView.addItemDecoration(decoration);
 
         return view;
