+++ app/src/main/java/com/yuganshtyagi/whatsapp_copy/Fragment/StatusFragment.java
@@ -1,44 +1,42 @@
-package com.yuganshtyagi.whatsapp_copy;
+package com.yuganshtyagi.whatsapp_copy.Fragment;
 
-import android.content.Context;
-import android.graphics.Canvas;
 import android.graphics.Color;
-import android.graphics.Paint;
-import android.graphics.Rect;
 import android.os.Bundle;
-import android.support.annotation.ColorInt;
-import android.support.annotation.FloatRange;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
-import android.util.TypedValue;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import com.yuganshtyagi.whatsapp_copy.Adapter.StatusAdapter;
+import com.yuganshtyagi.whatsapp_copy.R;
+import com.yuganshtyagi.whatsapp_copy.Utils.SeparatorDecoration;
 
 /**
  * Created by Yugansh Tyagi on 3/21/2018.
  */
 
-public class ChatsFragment extends Fragment {
+public class StatusFragment extends Fragment {
 
     RecyclerView recyclerView;
 
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.chats_fragment_layout, container, false);
+        View view = inflater.inflate(R.layout.status_fragment_layout, null, false);
+
 
         SeparatorDecoration decoration = new SeparatorDecoration(
                 getContext(),
                 Color.parseColor("#EAEAEA"),
                 0.5f);
-        recyclerView = view.findViewById(R.id.chat_rv);
+        recyclerView = view.findViewById(R.id.status_rv);
         recyclerView.setHasFixedSize(true);
         recyclerView.setLayoutManager(new LinearLayoutManager(getContext()));
-        recyclerView.setAdapter(new ChatAdapter(getContext()));
+        recyclerView.setAdapter(new StatusAdapter(getContext()));
         recyclerView.addItemDecoration(decoration);
 
         return view;
