+++ app/src/main/res/layout/camera_fragment_layout.xml
@@ -2,6 +2,7 @@
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:fitsSystemWindows="true"
     android:orientation="vertical">
 
@@ -10,12 +11,32 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent">
 
-        <TextView
+        <!--<TextView-->
+            <!--android:layout_width="match_parent"-->
+            <!--android:layout_height="match_parent"-->
+            <!--android:text="Camera preview will be Here"-->
+            <!--android:textSize="32sp"-->
+            <!--android:gravity="center"/>-->
+        <com.otaliastudios.cameraview.CameraView
+            android:id="@+id/camera_view_tool"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:text="Camera preview will be Here"
-            android:textSize="32sp"
-            android:gravity="center"/>
+            android:keepScreenOn="true"
+            app:cameraFacing="back"
+            app:cameraFlash="off"
+            app:cameraGrid="off"
+            app:cameraSessionType="picture"
+            app:cameraCropOutput="false"
+            app:cameraJpegQuality="100"
+            app:cameraVideoCodec="deviceDefault"
+            app:cameraWhiteBalance="auto"
+            app:cameraHdr="off"
+            app:cameraAudio="on"
+            app:cameraPlaySounds="true"
+            app:cameraVideoMaxSize="0"
+            app:cameraVideoMaxDuration="0"
+
+            />
 
     </FrameLayout>
